+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/action/ActionGraphSource.java
@@ -195,4 +195,8 @@ public class ActionGraphSource extends GraphSource {
         return ret;
     }
 
+    public HashMap<String, GraphTargetItem> getVariables() {
+        return variables;
+    }
+
 }